esptool/requirements.txt
bitstring==3.1.9
cffi==1.14.6
cryptography==3.4.8
ecdsa==0.17.0
esptool==3.1
pycparser==2.20
pyserial==3.5
reedsolo==1.5.4
six==1.16.0
